♻️ `repository.ts`のコードをDomainごとに分割
